use AgentCore runtime instead of Lambda (#5)
*Issue #, if available:*

closes #4 

*Description of changes:*

We replace Lambda with AgentCore Runtime, which is now GA and widely
available in many regions. By this replacement, we can now continue a
session up to 8 hours!

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.